✨カメラもウィンドウごとに作成 GameObjectにコピーコンストラクタを定義した。EntityIDは新しく生成して、statusとcomponentFlagだけコピーした

+++ mtgb/InputData.cpp
@@ -23,12 +23,12 @@ const bool mtgb::InputUtil::GetKeyDown(const KeyCode _keyCode, WindowContext _co
 		const InputData& inputFirstWnd = GetInput(WindowContext::First);
 		const InputData& inputSecondWnd = GetInput(WindowContext::Second);
 
-		return static_cast<bool>(KeyXOR(_keyCode, inputFirstWnd.keyStateCurrent_, inputFirstWnd.keyStatePrevious_) & inputFirstWnd.keyStateCurrent_[Index(_keyCode)])
-			|| static_cast<bool>(KeyXOR(_keyCode, inputSecondWnd.keyStateCurrent_, inputSecondWnd.keyStatePrevious_) & inputSecondWnd.keyStateCurrent_[Index(_keyCode)]);
+		return static_cast<bool>(KeyXOR(_keyCode, inputFirstWnd.keyStateCurrent_, inputFirstWnd.keyStatePrevious_) & static_cast<int>(inputFirstWnd.keyStateCurrent_[Index(_keyCode)]))
+			|| static_cast<bool>(KeyXOR(_keyCode, inputSecondWnd.keyStateCurrent_, inputSecondWnd.keyStatePrevious_) & static_cast<int>(inputSecondWnd.keyStateCurrent_[Index(_keyCode)]));
 	}
 
 	const InputData& input = GetInput(_context);
-	return static_cast<bool>(KeyXOR(_keyCode, input.keyStateCurrent_, input.keyStatePrevious_) & input.keyStateCurrent_[Index(_keyCode)]);
+	return static_cast<bool>(KeyXOR(_keyCode, input.keyStateCurrent_, input.keyStatePrevious_) & static_cast<int>(input.keyStateCurrent_[Index(_keyCode)]));
 }
 const bool mtgb::InputUtil::GetKeyUp(const KeyCode _keyCode, WindowContext _context)
 {
@@ -37,8 +37,8 @@ const bool mtgb::InputUtil::GetKeyUp(const KeyCode _keyCode, WindowContext _cont
 		const InputData& inputFirstWnd = GetInput(WindowContext::First);
 		const InputData& inputSecondWnd = GetInput(WindowContext::Second);
 
-		return static_cast<bool>(KeyXOR(_keyCode, inputFirstWnd.keyStateCurrent_, inputFirstWnd.keyStatePrevious_) & inputFirstWnd.keyStatePrevious_[Index(_keyCode)])
-			|| static_cast<bool>(KeyXOR(_keyCode, inputSecondWnd.keyStateCurrent_, inputSecondWnd.keyStatePrevious_) & inputSecondWnd.keyStatePrevious_[Index(_keyCode)]);
+		return static_cast<bool>(KeyXOR(_keyCode, inputFirstWnd.keyStateCurrent_, inputFirstWnd.keyStatePrevious_) & static_cast<int>(inputFirstWnd.keyStatePrevious_[Index(_keyCode)]))
+			|| static_cast<bool>(KeyXOR(_keyCode, inputSecondWnd.keyStateCurrent_, inputSecondWnd.keyStatePrevious_) & static_cast<int>(inputSecondWnd.keyStatePrevious_[Index(_keyCode)]));
 	}
 
 	const InputData& input = GetInput(_context);
